requirements.txt
asgiref==3.8.1
Django==4.2
django-jalali==7.2.0
django-jalali-date==1.1.3
jalali_core==1.0.0
jdatetime==5.0.0
persiantools==5.1.0
pillow==11.0.0
pytz==2024.2
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
mysqlclient